+++ CHANGELOG.md
@@ -1,6 +1,31 @@
 # CHANGELOG
 
+## v0.2.0
+
+- **Breaking**: Removes the use of Request option `:auth` to set WebSocket connect params. Use the `:connect_params` option to `AbsintheClient.WebSocket.connect/2` instead.
+
+For example, if you had this:
+
+```elixir
+req = Req.new(base_url: "http://localhost:4000/", auth: {:bearer, "valid-token"}) |> AbsintheClient.attach()
+ws = req |> AbsintheClient.WebSocket.connect!(url: "/socket/websocket")
+```
+
+...change it to this:
+
+```elixir
+req = Req.new(base_url: "http://localhost:4002/") |> AbsintheClient.attach()
+
+ws =
+  req
+  |> AbsintheClient.WebSocket.connect!(
+    url: "/socket/websocket",
+    connect_params: %{"token" => "valid-token"}
+  )
+```
+
+- The `:connect_params` option accepts a 0- or 1-arity function that returns a map of connect params.
+
 ## v0.1.0 (2022-10-05)
 
 Initial release
-